# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduFlow LMS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  image?: string | null;
}

export interface AuthTutor {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  profession?: string | null;
  /** @nullable */
  image?: string | null;
}

export interface AdminStats {
  totalStudents: number;
  totalPlaylists: number;
  totalVideos: number;
  totalComments: number;
  totalLikes?: number;
}

export interface TutorProfileUpdate {
  name?: string;
  profession?: string;
  image?: string;
  currentPassword?: string;
  newPassword?: string;
}

export type PlaylistInputStatus = typeof PlaylistInputStatus[keyof typeof PlaylistInputStatus];


export const PlaylistInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface PlaylistInput {
  title: string;
  description: string;
  image?: string;
  status?: PlaylistInputStatus;
}

export type PlaylistUpdateStatus = typeof PlaylistUpdateStatus[keyof typeof PlaylistUpdateStatus];


export const PlaylistUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface PlaylistUpdate {
  title?: string;
  description?: string;
  image?: string;
  status?: PlaylistUpdateStatus;
}

export interface Playlist {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  image?: string | null;
  status: string;
  tutorId?: number;
  createdAt?: string;
}

export interface PlaylistSummary {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  image?: string | null;
  status: string;
  tutorId?: number;
  /** @nullable */
  tutorName?: string | null;
  videoCount?: number;
  createdAt?: string;
}

export interface ContentItem {
  id: number;
  playlistId?: number;
  tutorId?: number;
  title: string;
  description: string;
  videoUrl: string;
  status: string;
  likeCount?: number;
  commentCount?: number;
  createdAt?: string;
}

export interface PlaylistDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  image?: string | null;
  status: string;
  tutorId?: number;
  /** @nullable */
  tutorName?: string | null;
  /** @nullable */
  tutorProfession?: string | null;
  videoCount?: number;
  createdAt?: string;
  videos: ContentItem[];
}

export type ContentInputStatus = typeof ContentInputStatus[keyof typeof ContentInputStatus];


export const ContentInputStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ContentInput {
  playlistId: number;
  title: string;
  description: string;
  videoUrl: string;
  status?: ContentInputStatus;
}

export type ContentUpdateStatus = typeof ContentUpdateStatus[keyof typeof ContentUpdateStatus];


export const ContentUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ContentUpdate {
  playlistId?: number;
  title?: string;
  description?: string;
  videoUrl?: string;
  status?: ContentUpdateStatus;
}

export interface ContentDetail {
  id: number;
  playlistId?: number;
  /** @nullable */
  playlistTitle?: string | null;
  tutorId?: number;
  /** @nullable */
  tutorName?: string | null;
  title: string;
  description: string;
  videoUrl: string;
  status: string;
  likeCount?: number;
  commentCount?: number;
  createdAt?: string;
}

export interface CommentInput {
  /** @minLength 1 */
  comment: string;
}

export interface Comment {
  id: number;
  contentId?: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  tutorId?: number | null;
  /** @nullable */
  userName?: string | null;
  comment: string;
  createdAt?: string;
}

export interface CommentWithContext {
  id: number;
  contentId?: number;
  /** @nullable */
  contentTitle?: string | null;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  tutorId?: number | null;
  /** @nullable */
  userName?: string | null;
  comment: string;
  createdAt?: string;
}

export interface LikeStatus {
  liked: boolean;
  count: number;
}

export interface BookmarkInput {
  playlistId: number;
}

export interface BookmarkStatus {
  bookmarked: boolean;
}

export interface BookmarkedPlaylist {
  bookmarkId?: number;
  playlistId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  tutorName?: string | null;
  videoCount?: number;
  bookmarkedAt?: string;
}

export interface Tutor {
  id: number;
  name: string;
  email?: string;
  /** @nullable */
  profession?: string | null;
  /** @nullable */
  image?: string | null;
  playlistCount?: number;
}

export interface TutorDetail {
  id: number;
  name: string;
  email?: string;
  /** @nullable */
  profession?: string | null;
  /** @nullable */
  image?: string | null;
  playlistCount?: number;
  playlists?: PlaylistSummary[];
}

export type ListPlaylistsParams = {
search?: string;
};

export type ListContentParams = {
playlistId?: number;
};