# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduFlow LMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerStudent
      tags: [auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginStudent
      tags: [auth]
      summary: Student login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentStudent
      tags: [auth]
      summary: Get current logged-in student
      responses:
        "200":
          description: Current student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Auth ──────────────────────────────────────────────────────────────
  /admin/login:
    post:
      operationId: loginAdmin
      tags: [admin]
      summary: Tutor/admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTutor"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current logged-in tutor
      responses:
        "200":
          description: Current tutor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTutor"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/profile:
    patch:
      operationId: updateAdminProfile
      tags: [admin]
      summary: Update tutor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTutor"
...
  /content/{id}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment to a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ── Likes ────────────────────────────────────────────────────────────────────
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: Get current user's bookmarks
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookmarkedPlaylist"
    post:
      operationId: addBookmark
      tags: [bookmarks]
      summary: Bookmark a playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bookmarks/{playlistId}:
    delete:
      operationId: removeBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bookmarks/{playlistId}/status:
    get:
      operationId: getBookmarkStatus
      tags: [bookmarks]
      summary: Check if a playlist is bookmarked
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bookmark status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkStatus"

  # ── Tutors ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Auth ─────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthUser:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        image:
          type: ["string", "null"]
    AuthTutor:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        profession:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
...
    TutorProfileUpdate:
      type: object
      properties:
        name:
          type: string
        profession:
          type: string
        image:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string

    # ── Playlists ─────────────────────────────────────────────────────────────
...
    CommentInput:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
          minLength: 1
    Comment:
      type: object
      required: [id, comment]
      properties:
        id:
          type: integer
        contentId:
          type: integer
        userId:
          type: ["integer", "null"]
        tutorId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
...
    BookmarkInput:
      type: object
      required: [playlistId]
      properties:
        playlistId:
          type: integer
    BookmarkStatus:
      type: object
      required: [bookmarked]
      properties:
        bookmarked:
          type: boolean
    BookmarkedPlaylist:
      type: object
      required: [playlistId, title]
      properties:
        bookmarkId:
          type: integer
        playlistId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        tutorName:
          type: ["string", "null"]
        videoCount:
          type: integer
        bookmarkedAt:
          type: string
          format: date-time

    # ── Tutors ────────────────────────────────────────────────────────────────
...