# lib/api-zod/src/generated/types/authTutor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduFlow LMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthTutor {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  profession?: string | null;
  /** @nullable */
  image?: string | null;
}