# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduFlow LMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  contentId?: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  tutorId?: number | null;
  /** @nullable */
  userName?: string | null;
  comment: string;
  createdAt?: Date;
}